+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * replaced custom simple UserManagement with proper Keycloak integration 
 * added `url` field to container image description
 * added `defaultValue` to action parameter description
+* added `includeConnected` parameter to `GET /containers` route, analogous to `GET /agents` and others
 * compatibility with Java 25